+++ src/main/java/co/kr/jurumarble/vote/controller/VoteController.java
@@ -108,14 +108,14 @@ public class VoteController {
         return new ResponseEntity(new GetVoteRecommendListResponse(voteRecommendListData), HttpStatus.OK);
     }
 
-    @Operation(summary = "투표 북마크", description = "헤더에 토큰 담고, 파라미터에 voteId 보내주시면 됩니다.")
-    @PostMapping("/{voteId}/bookmark")
-    public ResponseEntity bookmarkVote(@PathVariable Long voteId, @RequestAttribute Long userId) {
-
-        voteService.bookmarkVote(userId, voteId);
-
-        return new ResponseEntity(HttpStatus.OK);
-    }
+//    @Operation(summary = "투표 북마크", description = "헤더에 토큰 담고, 파라미터에 voteId 보내주시면 됩니다.")
+//    @PostMapping("/{voteId}/bookmark")
+//    public ResponseEntity bookmarkVote(@PathVariable Long voteId, @RequestAttribute Long userId) {
+//
+//        voteService.bookmarkVote(userId, voteId);
+//
+//        return new ResponseEntity(HttpStatus.OK);
+//    }
 
     @Operation(summary = "투표 참여 여부 조회", description = "파라미터에 voteId, 헤더에 userId 보내주시면 됩니다.")
     @GetMapping("/{voteId}/voted")
@@ -124,12 +124,12 @@ public class VoteController {
         return new ResponseEntity(new GetIsUserVotedResponse(userVoted),HttpStatus.OK);
     }
 
-    @Operation(summary = "북마크 여부 조회", description = "파라미어테 voteId, 헤더에 userId 보내주시면 됩니다.")
-    @GetMapping("/{voteId}/bookmark")
-    public ResponseEntity checkBookmarked(@PathVariable Long voteId, @RequestAttribute Long userId){
-
-        boolean result = voteService.checkBookmarked(userId, voteId);
-
-        return new ResponseEntity<>(new GetBookmarkedResponse(result), HttpStatus.OK);
-    }
+//    @Operation(summary = "북마크 여부 조회", description = "파라미어테 voteId, 헤더에 userId 보내주시면 됩니다.")
+//    @GetMapping("/{voteId}/bookmark")
+//    public ResponseEntity checkBookmarked(@PathVariable Long voteId, @RequestAttribute Long userId){
+//
+//        boolean result = voteService.checkBookmarked(userId, voteId);
+//
+//        return new ResponseEntity<>(new GetBookmarkedResponse(result), HttpStatus.OK);
+//    }
 }
