+++ src/test/java/co/kr/jurumarble/vote/service/VoteServiceTest.java
@@ -7,7 +7,7 @@ import co.kr.jurumarble.vote.domain.VoteContent;
 import co.kr.jurumarble.vote.domain.VoteDrinkContent;
 import co.kr.jurumarble.vote.domain.VoteGenerator;
 import co.kr.jurumarble.vote.enums.VoteType;
-import co.kr.jurumarble.vote.repository.BookmarkRepository;
+import co.kr.jurumarble.bookmark.repository.BookmarkRepository;
 import co.kr.jurumarble.vote.repository.VoteRepository;
 import co.kr.jurumarble.vote.repository.VoteResultRepository;
 import co.kr.jurumarble.vote.service.request.CreateDrinkVoteServiceRequest;
